+++ CHANGES.md
@@ -6,6 +6,10 @@ Version 0.2.0
 
 To be released.
 
+ -  Added two-factor authentication support.  [[#38]]
+
+[#38]: https://github.com/dahlia/hollo/issues/38
+
 
 Version 0.1.0
 -------------